% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_image.R
\name{download_image}
\alias{download_image}
\title{Download a map's image and/or thumbnail}
\usage{
download_image(id, position = 0)
}
\arguments{
\item{id}{Integer: The MapId of the entry to get an image from}

\item{position}{Numeric vector: The 0-based index of the image. Default is 0.
Can be a vector of multiple image indices}
}
\value{
Downloads map image png to working directory
}
\description{
Download a map's image and/or thumbnail
}
\examples{
# Download the image from Training - 01
download_image(id = 1984)
}
