% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_map.R
\name{download_map}
\alias{download_map}
\title{Downloads a map}
\usage{
download_map(ids)
}
\arguments{
\item{ids}{Numeric vector: The MapId of the map. Can be a vector of multiple
MapIds}
}
\value{
Downloads map to working directory
}
\description{
Downloads a map
}
\examples{
# Download Training - 01
download_map(ids = 1984)
}
