% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_name.R
\name{get_map_name}
\alias{get_map_name}
\title{Get the name of a map from a MapId}
\usage{
get_map_name(ids)
}
\arguments{
\item{ids}{Numeric vector: The MapId of the map. Can be a vector of multiple
MapIds}
}
\value{
A character vector containing the map name(s)
}
\description{
Get the name of a map from a MapId
}
\examples{
# Get the the map name from id 1984
get_map_name(ids = 1984)
}
