% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maptypes.R
\name{get_maptypes}
\alias{get_maptypes}
\title{Get all map types}
\usage{
get_maptypes()
}
\value{
A character vector of maptypes in descending order of occurrence
  count
}
\description{
Get all map types
}
