% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_titlepacks.R
\name{get_titlepacks}
\alias{get_titlepacks}
\title{Get all titlepacks}
\usage{
get_titlepacks()
}
\value{
A character vector of titlepacks in descending order of
  occurrence count
}
\description{
Get all titlepacks
}
