% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vehicles.R
\name{get_vehicles}
\alias{get_vehicles}
\title{Get all vehicles}
\usage{
get_vehicles()
}
\value{
A character list of vehicles. The order starts with the official
  vehicles (that have their own environment) by when they were introduced
  and afterwards the custom vehicles in descending order of occurrence count
}
\description{
Get all vehicles
}
