% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_maps.R
\name{search_maps}
\alias{search_maps}
\title{Search maps}
\source{
\url{https://api2.mania.exchange/search?version=2&s=2&q=Search\%20Maps}
}
\usage{
search_maps(params, fields)
}
\arguments{
\item{params}{A named list of input parameters. The full list of parameters
can be seen in the source section}

\item{fields}{A string of output parameters. The full list of parameters
can be seen in the source section. See the example for how this should
be structured}
}
\value{
A dataframe with columns being the output parameters listed in
  the fields argument
}
\description{
Search maps
}
\details{
If no fields are provided, the default fields in the output are
  Name, MapId, Authors, MapType, Difficulty, AwardCount, UploadedAt,
  OnlineWR, DownloadCount, and Tags
}
\examples{
# Search 50 Ubisoft Nadeo maps
params <- list(author = "Ubisoft Nadeo", count = 50)
fields <- "Name,MapId,Authors[],AwardCount,Difficulty"
search_maps(params = params, fields = fields)
}
